refact: add base class
