+++ lib/dynamodb/api.rb
@@ -5,6 +5,7 @@ require 'active_support/core_ext'
 require 'dynamodb/api/version'
 require 'dynamodb/api/config'
 require 'dynamodb/api/adapter'
+require 'dynamodb/api/base'
 require 'dynamodb/api/query'
 require 'dynamodb/api/relation'
 require 'dynamodb/api/relation/query_methods'
